Prac2/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.5
django-appconf==1.1.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-extensions==4.1
django-fernet-encrypted-fields==0.3.0
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
idna==3.10
phonenumbers==9.0.13
pillow==11.3.0
pycparser==2.22
pypng==0.20220715.0
python-magic==0.4.27
qrcode==7.4.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
